Moved attack funcs to separate folders
