+++ attacks/collusion/__init__.py
@@ -1,5 +1,1 @@
-from attacks.collusion.delta_to_zero import DeltaToZeroAttack
-from attacks.collusion.paf import ScalarPAFAttack
-
-
-ATTACKS = {'delta_to_zero': DeltaToZeroAttack, 'paf': ScalarPAFAttack}
+from attacks.collusion.attack_funcs import ScalarPAFAttack, DeltaToZeroAttack, COLLUSION_ATTACKS
