+++ attacks/collusion/attack_funcs/__init__.py
@@ -1,0 +1,5 @@
+from attacks.collusion.attack_funcs.delta_to_zero import DeltaToZeroAttack
+from attacks.collusion.attack_funcs.paf import ScalarPAFAttack
+
+
+COLLUSION_ATTACKS = {'delta_to_zero': DeltaToZeroAttack, 'paf': ScalarPAFAttack}
