+++ attacks/collusion/attack_funcs/base.py
