+++ attacks/local/attack_funcs/base.py
