+++ attacks/local/attack_funcs/gaussian.py
@@ -1,6 +1,6 @@
 import random
 import tensorflow as tf
-from attacks.local.base import LocalAttack
+from attacks.local.attack_funcs.base import LocalAttack
 
 
 class GaussianAttack(LocalAttack):
